cdd-admin-srv/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-cors-middleware==1.3.1
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.3
pytz==2018.4
raven==6.9.0
redis==2.10.6
requests==2.18.4
schedule==0.5.0
six==1.11.0
urllib3==1.22
